+++ anton/core/tools/tool_defs.py
@@ -1,0 +1,145 @@
+SCRATCHPAD_TOOL = {
+    "name": "scratchpad",
+    "description": (
+        "Run Python code in a persistent scratchpad. Use this whenever you need to "
+        "count characters, do math, parse data, transform text, or any task that "
+        "benefits from precise computation rather than guessing. Variables, imports, "
+        "and data persist across cells — like a notebook you drive programmatically.\n\n"
+        "Actions:\n"
+        "- exec: Run code in the scratchpad (creates it if needed)\n"
+        "- view: See all cells and their outputs\n"
+        "- reset: Restart the process, clearing all state (installed packages survive)\n"
+        "- remove: Kill the scratchpad and delete its environment\n"
+        "- dump: Show a clean notebook-style summary of cells (code + truncated output)\n"
+        "- install: Install Python packages into the scratchpad's environment. "
+        "Packages persist across resets.\n\n"
+        "Use print() to produce output. Host Python packages are available by default. "
+        "Include a 'packages' array on exec calls for any libraries your code needs — "
+        "they'll be auto-installed before the cell runs (already-installed ones are skipped).\n"
+        "get_llm() returns a pre-configured LLM client (sync) — call "
+        "llm.complete(system=..., messages=[...]) for AI-powered computation.\n"
+        "llm.generate_object(MyModel, system=..., messages=[...]) extracts structured "
+        "data into Pydantic models. Supports single models and list[Model].\n"
+        "agentic_loop(system=..., user_message=..., tools=[...], handle_tool=fn) "
+        "runs a tool-call loop where the LLM reasons and calls your tools iteratively. "
+        "handle_tool(name, inputs) -> str is a plain sync function.\n"
+        "sample(var) inspects any variable with type-aware formatting — DataFrames get "
+        "shape/dtypes/head, dicts get keys/values, lists get length/items. "
+        "Defaults to 'preview' mode (compact); use sample(var, mode='full') for complete dump.\n"
+        "All .anton/.env secrets are available as environment variables (os.environ).\n\n"
+        "IMPORTANT: Cells have an inactivity timeout of 30 seconds — if a cell produces "
+        "no output and no progress() calls for 30s, it is killed and all state is lost. "
+        "For long-running code (API calls, data extraction, heavy computation), call "
+        "progress(message) periodically to signal work is ongoing and reset the timer. "
+        "The total timeout scales from your estimated_execution_time_seconds "
+        "(roughly 2x the estimate). You MUST provide estimated_execution_time_seconds "
+        "for every exec call. For very long operations, provide a realistic estimate "
+        "and use progress() to keep the cell alive."
+    ),
+    "input_schema": {
+        "type": "object",
+        "properties": {
+            "action": {"type": "string", "enum": ["exec", "view", "reset", "remove", "dump", "install"]},
+            "name": {"type": "string", "description": "Scratchpad name"},
+            "code": {
+                "type": "string",
+                "description": "Python code (exec only). Use print() for output.",
+            },
+            "packages": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": "Package names needed by this cell (exec or install). "
+                "Listed after code so you know exactly what to include. "
+                "Already-installed packages are skipped automatically.",
+            },
+            "one_line_description": {
+                "type": "string",
+                "description": "Brief description of what this cell does (e.g. 'Scrape listing prices'). Required for exec.",
+            },
+            "estimated_execution_time_seconds": {
+                "type": "integer",
+                "description": "Estimated execution time in seconds. Drives the total timeout (roughly 2x estimate). Use progress() for long cells.",
+            },
+        },
+        "required": ["action", "name"],
+    },
+}
+
+
+MEMORIZE_TOOL = {
+    "name": "memorize",
+    "description": (
+        "Encode a rule or lesson into long-term memory for future sessions. "
+        "Use this when you learn something important, discover a useful pattern, "
+        "or the user asks you to remember something.\n\n"
+        "Entry kinds:\n"
+        "- always: Something to always do ('Use httpx instead of requests')\n"
+        "- never: Something to never do ('Never use time.sleep() in scratchpad')\n"
+        "- when: Conditional rule ('If paginated API → use async + progress()')\n"
+        "- lesson: Factual knowledge ('CoinGecko rate-limits at 50/min')\n"
+        "- profile: Fact about the user ('Name: Jorge', 'Prefers dark mode')"
+    ),
+    "input_schema": {
+        "type": "object",
+        "properties": {
+            "entries": {
+                "type": "array",
+                "items": {
+                    "type": "object",
+                    "properties": {
+                        "text": {
+                            "type": "string",
+                            "description": "The memory to encode",
+                        },
+                        "kind": {
+                            "type": "string",
+                            "enum": ["always", "never", "when", "lesson", "profile"],
+                        },
+                        "scope": {
+                            "type": "string",
+                            "enum": ["global", "project"],
+                        },
+                        "topic": {
+                            "type": "string",
+                            "description": "Topic slug for lessons (e.g. 'api-coingecko')",
+                        },
+                    },
+                    "required": ["text", "kind", "scope"],
+                },
+            },
+        },
+        "required": ["entries"],
+    },
+}
+
+
+RECALL_TOOL = {
+    "name": "recall",
+    "description": (
+        "Search your episodic memory — an archive of past conversations. "
+        "ONLY use this when the user explicitly asks about a previous conversation "
+        "or session (e.g. 'what did we talk about last time?', 'remember when we...', "
+        "'have we discussed X before?'). Do NOT use this for questions about code, "
+        "files, or data in the workspace — use the scratchpad to explore those directly.\n\n"
+        "Returns timestamped episodes matching the query (newest first). "
+        "A single call is enough — do not call multiple times with different queries."
+    ),
+    "input_schema": {
+        "type": "object",
+        "properties": {
+            "query": {
+                "type": "string",
+                "description": "Search term to find in past conversations.",
+            },
+            "max_results": {
+                "type": "integer",
+                "description": "Maximum episodes to return (default 20).",
+            },
+            "days_back": {
+                "type": "integer",
+                "description": "Only search episodes from the last N days.",
+            },
+        },
+        "required": ["query"],
+    },
+}
