+++ anton/core/tools/tool_handlers.py
@@ -1,0 +1,196 @@
+from anton.core.session import ChatSession
+
+
+async def handle_recall(session: ChatSession, tc_input: dict) -> str:
+    """Process a recall tool call — search episodic memory."""
+    if session._episodic is None or not session._episodic.enabled:
+        return "Episodic memory is not available."
+
+    query = tc_input.get("query", "")
+    if not query:
+        return "No query provided."
+
+    kwargs: dict = {}
+    if "max_results" in tc_input:
+        kwargs["max_results"] = int(tc_input["max_results"])
+    if "days_back" in tc_input:
+        kwargs["days_back"] = int(tc_input["days_back"])
+
+    return session._episodic.recall_formatted(query, **kwargs)
+
+
+async def handle_memorize(session: ChatSession, tc_input: dict) -> str:
+    """Process a memorize tool call and return a result string.
+
+    Encoding is fire-and-forget so it never blocks scratchpad execution.
+    """
+    import asyncio
+
+    if session._cortex is None:
+        return "Memory system not available."
+
+    if session._cortex.mode == "off":
+        return "Memory encoding is disabled. Change memory mode via /setup to enable."
+
+    from anton.memory.hippocampus import Engram
+
+    raw_entries = tc_input.get("entries", [])
+    if not raw_entries:
+        return "No entries provided."
+
+    engrams: list[Engram] = []
+    for entry in raw_entries:
+        if not isinstance(entry, dict) or "text" not in entry:
+            continue
+
+        kind = entry.get("kind", "lesson")
+        if kind not in ("always", "never", "when", "lesson", "profile"):
+            kind = "lesson"
+
+        scope = entry.get("scope", "project")
+        if scope not in ("global", "project"):
+            scope = "project"
+
+        # User-sourced memories (via explicit tool call) get high confidence
+        engrams.append(Engram(
+            text=entry["text"],
+            kind=kind,
+            scope=scope,
+            confidence="high",
+            topic=entry.get("topic", ""),
+            source="user",
+        ))
+
+    if not engrams:
+        return "No valid entries provided."
+
+    # Always encode immediately via fire-and-forget — the LLM explicitly
+    # chose to memorize these, so we never interrupt the user mid-turn
+    # with confirmation prompts.  Confirmations are reserved for the
+    # post-turn consolidator (lessons extracted from scratchpad sessions).
+    async def _encode_bg(cortex, entries):
+        try:
+            await cortex.encode(entries)
+        except Exception:
+            pass  # Best-effort; don't disrupt the conversation
+
+    asyncio.create_task(_encode_bg(session._cortex, engrams))
+
+    descriptions = [f"Encoded {e.kind}: {e.text}" for e in engrams]
+    return "Memory updated: " + "; ".join(descriptions)
+
+
+async def prepare_scratchpad_exec(session: ChatSession, tc_input: dict):
+    """Validate and prepare a scratchpad exec call.
+
+    Returns (pad, code, description, estimated_time, estimated_seconds) or
+    a str error message if validation fails.
+    """
+    name = tc_input.get("name", "")
+    code = tc_input.get("code", "")
+    if not code or not code.strip():
+        return "No code provided."
+
+    pad = await session._scratchpads.get_or_create(name)
+
+    # Auto-install packages before running the cell
+    packages = tc_input.get("packages", [])
+    if packages:
+        install_result = await pad.install_packages(packages)
+        if "Install failed" in install_result or "timed out" in install_result:
+            return install_result
+
+    description = tc_input.get("one_line_description", "")
+    estimated_seconds = tc_input.get("estimated_execution_time_seconds", 0)
+    if isinstance(estimated_seconds, str):
+        try:
+            estimated_seconds = int(estimated_seconds)
+        except ValueError:
+            estimated_seconds = 0
+
+    estimated_time = f"{estimated_seconds}s" if estimated_seconds > 0 else ""
+    return pad, code, description, estimated_time, estimated_seconds
+
+
+def format_cell_result(cell) -> str:
+    """Format a Cell into a tool result string.
+
+    Every section is labeled so the LLM can tell what came from where:
+    [output] — print() / stdout from the cell code
+    [logs]   — library logging (httpx, urllib3, etc.) captured at INFO+
+    [stderr] — warnings and stderr writes
+    [error]  — Python traceback if the cell raised an exception
+    """
+    parts: list[str] = []
+    if cell.stdout:
+        stdout = cell.stdout
+        if len(stdout) > 10_000:
+            stdout = stdout[:10_000] + f"\n\n... (truncated, {len(stdout)} chars total)"
+        parts.append(f"[output]\n{stdout}")
+    if cell.logs if hasattr(cell, "logs") else False:
+        logs = cell.logs.strip()
+        if len(logs) > 3_000:
+            logs = logs[:3_000] + "\n... (logs truncated)"
+        parts.append(f"[logs]\n{logs}")
+    if cell.stderr:
+        parts.append(f"[stderr]\n{cell.stderr}")
+    if cell.error:
+        parts.append(f"[error]\n{cell.error}")
+    if not parts:
+        return "Code executed successfully (no output)."
+    return "\n".join(parts)
+
+
+async def handle_scratchpad(session: ChatSession, tc_input: dict) -> str:
+    """Dispatch a scratchpad tool call by action."""
+    action = tc_input.get("action", "")
+    name = tc_input.get("name", "")
+
+    if not name:
+        return "Scratchpad name is required."
+
+    if action == "exec":
+        result = await prepare_scratchpad_exec(session, tc_input)
+        if isinstance(result, str):
+            return result
+        pad, code, description, estimated_time, estimated_seconds = result
+
+        cell = await pad.execute(
+            code,
+            description=description,
+            estimated_time=estimated_time,
+            estimated_seconds=estimated_seconds,
+        )
+        return format_cell_result(cell)
+
+    elif action == "view":
+        pad = session._scratchpads._pads.get(name)
+        if pad is None:
+            return f"No scratchpad named '{name}'."
+        return pad.view()
+
+    elif action == "reset":
+        pad = session._scratchpads._pads.get(name)
+        if pad is None:
+            return f"No scratchpad named '{name}'."
+        await pad.reset()
+        return f"Scratchpad '{name}' reset. All state cleared."
+
+    elif action == "remove":
+        return await session._scratchpads.remove(name)
+
+    elif action == "dump":
+        pad = session._scratchpads._pads.get(name)
+        if pad is None:
+            return f"No scratchpad named '{name}'."
+        return pad.render_notebook()
+
+    elif action == "install":
+        packages = tc_input.get("packages", [])
+        if not packages:
+            return "No packages specified."
+        pad = await session._scratchpads.get_or_create(name)
+        return await pad.install_packages(packages)
+
+    else:
+        return f"Unknown scratchpad action: {action}"
